Coin dispatch: removing UTXOPlanner duplication, merged into AnySigner.
