+++ codegen/lib/templates/swift/TrustWalletCore.h.erb
@@ -19,7 +19,6 @@ FOUNDATION_EXPORT const unsigned char TrustWalletCoreVersionString[];
 #include <TrustWalletCore/TWFoundationString.h>
 
 #include <TrustWalletCore/TWAnySigner.h>
-#include <TrustWalletCore/TWUTXOPlanner.h>
 
 <% entities.sort { |x,y| x.name <=> y.name }.each do |entity| -%>
 #include <TrustWalletCore/TW<%= entity.name %>.h>
