--- include/TrustWalletCore/TWUTXOPlanner.h
@@ -1,19 +1,0 @@
-// Copyright © 2017-2020 Trust Wallet.
-//
-// This file is part of Trust. The full Trust copyright notice, including
-// terms governing use, modification, and redistribution, is contained in the
-// file LICENSE at the root of the source code distribution tree.
-
-#pragma once
-
-#include "TWBase.h"
-#include "TWData.h"
-#include "TWCoinType.h"
-
-TW_EXTERN_C_BEGIN
-
-struct TWUTXOPlanner;
-
-TWData * _Nonnull TWUTXOPlannerPlan(TWData *_Nonnull input, enum TWCoinType coin);
-
-TW_EXTERN_C_END
