+++ jni/cpp/AnySigner.h
@@ -15,6 +15,9 @@ TW_EXTERN_C_BEGIN
 JNIEXPORT
 jbyteArray JNICALL Java_wallet_core_java_AnySigner_nativeSign(JNIEnv *env, jclass thisClass, jbyteArray input, jint coin);
 
+JNIEXPORT
+jbyteArray JNICALL Java_wallet_core_java_AnySigner_nativePlan(JNIEnv *env, jclass thisClass, jbyteArray input, jint coin);
+
 TW_EXTERN_C_END
 
 #endif // JNI_TW_ANYSIGNER_H
