--- jni/cpp/UTXOPlanner.c
@@ -1,20 +1,0 @@
-// Copyright © 2017-2020 Trust Wallet.
-//
-// This file is part of Trust. The full Trust copyright notice, including
-// terms governing use, modification, and redistribution, is contained in the
-// file LICENSE at the root of the source code distribution tree.
-
-#include <jni.h>
-#include <stdio.h>
-#include <string.h>
-
-#include "UTXOPlanner.h"
-#include "TWJNI.h"
-
-jbyteArray JNICALL Java_wallet_core_java_UTXOPlanner_nativePlan(JNIEnv *env, jclass thisClass, jbyteArray input, jint coin) {
-    TWData *inputData = TWDataCreateWithJByteArray(env, input);
-    TWData *outputData = TWUTXOPlannerPlan(inputData, coin);
-    jbyteArray resultData = TWDataJByteArray(outputData, env);
-    TWDataDelete(inputData);
-    return resultData;
-}
