--- src/interface/TWUTXOPlanner.cpp
@@ -1,18 +1,0 @@
-// Copyright © 2017-2020 Trust Wallet.
-//
-// This file is part of Trust. The full Trust copyright notice, including
-// terms governing use, modification, and redistribution, is contained in the
-// file LICENSE at the root of the source code distribution tree.
-
-#include <TrustWalletCore/TWUTXOPlanner.h>
-
-#include "Coin.h"
-
-using namespace TW;
-
-TWData* _Nonnull TWUTXOPlannerPlan(TWData* _Nonnull data, enum TWCoinType coin) {
-    const Data& dataIn = *(reinterpret_cast<const Data*>(data));
-    Data dataOut;
-    TW::anyCoinPlan(coin, dataIn, dataOut);
-    return TWDataCreateWithBytes(dataOut.data(), dataOut.size());
-}
