+++ tests/Decred/TWAnySignerTests.cpp
@@ -13,7 +13,6 @@
 #include <TrustWalletCore/TWBitcoinScript.h>
 #include <TrustWalletCore/TWBitcoinSigHashType.h>
 #include <TrustWalletCore/TWAnySigner.h>
-#include <TrustWalletCore/TWUTXOPlanner.h>
 #include <gtest/gtest.h>
