+++ tests/Monacoin/TWMonacoinTransactionTests.cpp
@@ -12,7 +12,6 @@
 #include <TrustWalletCore/TWBitcoinScript.h>
 #include <TrustWalletCore/TWBitcoinSigHashType.h>
 #include <TrustWalletCore/TWAnySigner.h>
-#include <TrustWalletCore/TWUTXOPlanner.h>
 
 #include <gtest/gtest.h>
